Further refactoring Fixed bug where players could not abaddon race with 0 speed Fixed bug with wrong color of text when choosing action Fixed bug where chance box was not correctly show estimated chance of succeeding Changed attack mechanics - now attacks have a drawback, attacks are more impactful in general

+++ Project3/Participant.h
@@ -47,7 +47,7 @@ public:
 	float current_durability;
 	SinglePlayer *network_role;
 	std::string car_path;	
-	int attacked;
+	float attacked;
 	bool drift;
 
 	Participant(std::string name, std::string car_path, std::string tire_path, SinglePlayer &network_role);	//Creating human player
